Estavel/old/Versao_Estavel_Funcional/requirements.txt
#requirements.txt
# === Gestão de variáveis de ambiente ===
python-dotenv

# === Extração de texto / PDF ===
pymongo
pytesseract
pdf2image
pdfminer.six
PyPDF2
PyMuPDF
unstructured
python-docx

# Novas ferramentas de extração
pdfplumber
tika
boto3

# === Chunking semântico / embeddings ===
langchain
langchain-community
sentence-transformers
torch
scikit-learn

# === Banco de dados ===
psycopg2-binary
requests

# === Outras utilidades ===
gridfs  # via pymongo

prometheus_client